update asus vts rule list

+++ src/routers.rs
@@ -8,5 +8,5 @@ pub trait Router {
     fn descriptor(&self) -> String;
     async fn probe(&mut self) -> Result<()>;
     async fn login(&mut self, credentials: Vec<String>) -> Result<()>;
-    async fn configure(&mut self, ip: IpAddr, ports: Vec<u8>) -> Result<()>;
+    async fn configure(&mut self, ip: IpAddr, ports: Vec<u16>) -> Result<()>;
 }
